+++ src/components/layouts/header/variants/header-back-button.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useRouter } from 'next/navigation';
-import HeaderLayer from './header-layer';
+import HeaderLayer from '../header-wrapper';
 
 const HeaderBackButton = () => {
   const router = useRouter();
