+++ src/components/layouts/header/variants/header-default.tsx
@@ -1,10 +1,8 @@
-'use client';
-
 import Image from 'next/image';
 import LOGO from '@/../public/logo.svg';
 import Link from 'next/link';
 import SITE_MAP from '@/constants/site-map.constant';
-import HeaderLayer from './header-layer';
+import HeaderLayer from '../header-wrapper';
 
 const Header = () => {
   return (
